MessageBuffer implemented as PushPullBuffer of received items. Each strategy receives only one type of data type and each TimerTaskListener sends only one data type. All objects are renamed according the new convention.

+++ jdeeco-gossip-plugin/src/cz/cuni/mff/d3s/jdeeco/gossip/GossipProperties.java
@@ -42,13 +42,13 @@ public final class GossipProperties {
 	/**
 	 * Default value of knowledge publishing period in milliseconds.
 	 */
-	public static final int KNOWLEDGE_PUSH_PERIOD_DEFAULT = 1000;
+	public static final long KNOWLEDGE_PUSH_PERIOD_DEFAULT = 1000;
 	/**
 	 * See {@link #KNOWLEDGE_PUSH_PERIOD_DEFAULT} for default value.
 	 * @return Knowledge publishing period in milliseconds.
 	 */
-	public static int getKnowledgePushPeriod() {
-		return Integer.getInteger(KNOWLEDGE_PUSH_PERIOD, KNOWLEDGE_PUSH_PERIOD_DEFAULT);
+	public static long getKnowledgePushPeriod() {
+		return Long.getLong(KNOWLEDGE_PUSH_PERIOD, KNOWLEDGE_PUSH_PERIOD_DEFAULT);
 	}
 	
 	public static final String KONWLEDGE_PUSH_PROBABILITY = "deeco.push.knowledge_probability";
@@ -61,13 +61,13 @@ public final class GossipProperties {
 	}
 	
 	public static final String HEADERS_PUSH_PERIOD = "deeco.push.headers_period";
-	public static final int HEADERS_PUSH_PERIOD_DEFAULT = 2000;
+	public static final long HEADERS_PUSH_PERIOD_DEFAULT = 2000;
 	/**
 	 * 
 	 * @return Message headers publishing period in milliseconds.
 	 */
-	public static int getHeadersPushPeriod() {
-		return Integer.getInteger(HEADERS_PUSH_PERIOD, HEADERS_PUSH_PERIOD_DEFAULT);
+	public static long getHeadersPushPeriod() {
+		return Long.getLong(HEADERS_PUSH_PERIOD, HEADERS_PUSH_PERIOD_DEFAULT);
 	}
 	
 	public static final String HEADERS_PUSH_PROBABILITY = "deeco.push.headers_probability";
@@ -87,24 +87,24 @@ public final class GossipProperties {
 	}*/
 	
 	public static final String KNOWLEDGE_PULL_PERIOD = "deeco.pull.knowledge_period";
-	public static final int KNOWLEDGE_PULL_PERIOD_DEFAULT = 4000;
+	public static final long KNOWLEDGE_PULL_PERIOD_DEFAULT = 4000;
 	/**
 	 * 
 	 * @return Knowledge pulling period in milliseconds.
 	 */
-	public static final int getKnowledgePullPeriod() {
-		return Integer.getInteger(KNOWLEDGE_PULL_PERIOD, KNOWLEDGE_PULL_PERIOD_DEFAULT);
+	public static final long getKnowledgePullPeriod() {
+		return Long.getLong(KNOWLEDGE_PULL_PERIOD, KNOWLEDGE_PULL_PERIOD_DEFAULT);
 	}
 	
 	public static final String KNOWLEDGE_PULL_TIMEOUT = "deeco.pull.knowledge_timeout";
-	public static final int KNOWLEDGE_PULL_TIMEOUT_DEFAULT = 4000;
+	public static final long KNOWLEDGE_PULL_TIMEOUT_DEFAULT = 4000;
 	/**
 	 * See {@link #KNOWLEDGE_PULL_TIMEOUT_DEFAULT} for default value.
 	 * @return Time after which knowledge is considered to be outdated and PULL request
 	 * is necessary.
 	 */
-	public static final int getKnowledgePullTimeout() {
-		return Integer.getInteger(KNOWLEDGE_PULL_TIMEOUT, KNOWLEDGE_PULL_TIMEOUT_DEFAULT);
+	public static final long getKnowledgePullTimeout() {
+		return Long.getLong(KNOWLEDGE_PULL_TIMEOUT, KNOWLEDGE_PULL_TIMEOUT_DEFAULT);
 	}
 	
 	public static final String COMPONENT_PULL_TIMEOUT = "deeco.pull.component_timeout";
@@ -114,8 +114,8 @@ public final class GossipProperties {
 	 * @return Time after which component is considered to be decommissioned and
 	 * can be removed from all buffers.
 	 */
-	public static final int getComponentPullTimeout() {
-		return Integer.getInteger(COMPONENT_PULL_TIMEOUT, COMPONENT_PULL_TIMEOUT_DEFAULT);
+	public static final long getComponentPullTimeout() {
+		return Long.getLong(COMPONENT_PULL_TIMEOUT, COMPONENT_PULL_TIMEOUT_DEFAULT);
 	}
 	
 	private static double getDouble(String name, double def) {
